pdfparser/src/requirements.txt
beautifulsoup4>=4.4.0
pdfminer.six>=20140328
PyPDF2>=1.25.1
requests>=2.7.0
pycti
PyYAML
pymisp
urllib3
python-dateutil
stix2
